back_end/requirements.txt
Flask==2.1.2
gunicorn==19.9.0
flask_cors
azure-messaging-webpubsubservice